requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
autopep8==2.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
fastapi==0.89.1
google-api-core==2.11.0
google-auth==2.16.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
pandas==1.5.3
phonenumbers==8.13.7
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
pymongo==4.3.3
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-http-client==3.3.7
python-multipart==0.0.6
pytz==2022.7.1
redis==4.4.2
requests==2.28.2
rsa==4.9
sendgrid==6.10.0
six==1.16.0
sniffio==1.3.0
starkbank-ecdsa==2.2.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.20.0
XlsxWriter==3.0.8
